Modularized utility functions and middlewares
