+++ libs/functions/getArgs.js
@@ -1,0 +1,12 @@
+function getArgs() {
+    const arguments = process.argv.slice(2);
+    let argDict = {};
+
+    arguments.forEach(arg => {
+        let keyValue = arg.split("=")
+        argDict[keyValue[0]] = keyValue[1];
+    });
+    return argDict;
+}
+
+module.exports = getArgs;
